refactor(be): move DKIM/DMARC orchestration into their modules
Sea-snake review feedback: the typestate module should orchestrate
across `dkim` and `dmarc` rather than reimplement them. Moves the
two helpers that were doing per-protocol work in
`email_recovery/typestate.rs` to their respective crates:

- `dmarc::compute_outcome(dkim_domain, from_domain, dmarc_txt) ->
  DmarcOutcome` lives in `dmarc/mod.rs` now (with its unit tests
  alongside).

- `dkim::run_signature_check(sig, headers, body, dkim_header_value,
  dkim_txt, now_secs) -> Result<Vec<DkimCheck>, …>` lives in
  `dkim/verify.rs` now. Takes primitive args (no typestate types),
  so the dkim crate has no upward dependency on email_recovery.

Stage 3's `verify_signature` collapses to ~20 lines of glue: call
`dkim::run_signature_check`, then `dmarc::extract_from_domain` for
the From-domain that DMARC alignment needs. Stage 3's `try_from`
calls `dmarc::compute_outcome` directly. The typestate file dropped
~160 lines.
